client: pull FetchAndParseJSON
